+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fix amount base of the payment. Now the base has the value of the shipping and the amount of the products.
+- Fix locale sent to request in WC
 
 ## [3.7.1 (2022-09-15)](https://github.com/placetopay/prestashop-placetopay/compare/3.7.0...3.7.1)